* xscandir.c - fix scandir() brokenness, just cast dselect to void so   that it will work no matter if scandir expects const or non-const * ircaux.[ch] - remove vsnprintf/snprintf - configure.ac wasn't even checking for   them and the overrides were silly and broken. * merged some freebsd ports patches

+++ iflist.c
@@ -184,7 +184,7 @@ int ifaces(iface_cb callback, void *data)
     c->thr_ifcb = callback;
     c->thr_data = data;
 
-    return THR_CREATE(ifaces_r, c);
+    return (int)THR_CREATE(ifaces_r, c);
 }
 
 /**
